docker/requirements.txt
# Dependencies necessary to execute run_docker.py
absl-py==0.13.0
docker==5.0.2